+++ packages/cosmos-ts/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.31](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/cosmos-ts@0.0.29...@injectivelabs/cosmos-ts@0.0.31) (2021-12-07)
+
+**Note:** Version bump only for package @injectivelabs/cosmos-ts
+
+
+
+
+
+## [0.0.30](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/cosmos-ts@0.0.29...@injectivelabs/cosmos-ts@0.0.30) (2021-12-07)
+
+**Note:** Version bump only for package @injectivelabs/cosmos-ts
+
+
+
+
+
 ## [0.0.29](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/cosmos-ts@0.0.28...@injectivelabs/cosmos-ts@0.0.29) (2021-12-07)